Add manager class
