+++ src/index.html
@@ -20,7 +20,7 @@
         <ul id="task-list">
         </ul>
         <div class="clear">
-          <button class="border-0 bg-transparent text-secondary" type="submit">Clear all completed</button>
+          <button class="border-0 bg-transparent text-secondary" id="add-button" type="submit">Clear all completed</button>
         </div>
       </div>
     </main>
